refactor(ui): refactor social signin callback flow
refactor social signin callback flow
