+++ packages/ui/src/App.tsx
@@ -14,6 +14,7 @@ import Passcode from './pages/Passcode';
 import Register from './pages/Register';
 import SecondarySignIn from './pages/SecondarySignIn';
 import SignIn from './pages/SignIn';
+import SocialSignInCallback from './pages/SignIn/SocialCallback';
 import SocialRegister from './pages/SocialRegister';
 import getSignInExperienceSettings from './utils/sign-in-experience';
 
@@ -54,7 +55,7 @@ const App = () => {
             <Route path="/" element={<Navigate replace to="/sign-in" />} />
             <Route path="/sign-in" element={<SignIn />} />
             <Route path="/sign-in/consent" element={<Consent />} />
-            <Route path="/sign-in/callback/:connector" element={<SignIn />} />
+            <Route path="/sign-in/callback/:connector" element={<SocialSignInCallback />} />
             <Route path="/sign-in/:method" element={<SecondarySignIn />} />
             <Route path="/register" element={<Register />} />
             <Route path="/register/:method" element={<Register />} />
