+++ packages/ui/src/containers/SocialSignIn/SecondarySocialSignIn/index.test.tsx
@@ -1,12 +1,11 @@
 import { fireEvent, waitFor } from '@testing-library/react';
 import React from 'react';
-import { MemoryRouter, Route, Routes } from 'react-router-dom';
+import { MemoryRouter } from 'react-router-dom';
 
 import renderWithPageContext from '@/__mocks__/RenderWithPageContext';
 import SettingsProvider from '@/__mocks__/RenderWithPageContext/SettingsProvider';
 import { socialConnectors, mockSignInExperienceSettings } from '@/__mocks__/logto';
 import * as socialSignInApi from '@/apis/social';
-import { generateState, storeState } from '@/hooks/utils';
 
 import SecondarySocialSignIn, { defaultSize } from '.';
 
@@ -15,10 +14,6 @@ describe('SecondarySocialSignIn', () => {
 
   const invokeSocialSignInSpy = jest
     .spyOn(socialSignInApi, 'invokeSocialSignIn')
-    .mockResolvedValue({ redirectTo: `${mockOrigin}/callback` });
-
-  const signInWithSocialSpy = jest
-    .spyOn(socialSignInApi, 'signInWithSocial')
     .mockResolvedValue({ redirectTo: `${mockOrigin}/callback` });
 
   beforeEach(() => {
@@ -125,35 +120,5 @@ describe('SecondarySocialSignIn', () => {
       expect(invokeSocialSignInSpy).toBeCalled();
       expect(logtoNativeSdk?.getPostMessage).toBeCalled();
     }
-  });
-
-  it('callback validation and signIn with social', async () => {
-    const state = generateState();
-    storeState(state, 'github');
-
-    /* eslint-disable @silverhand/fp/no-mutating-methods */
-    Object.defineProperty(window, 'location', {
-      value: {
-        href: `/sign-in/callback?state=${state}&code=foo`,
-        search: `?state=${state}&code=foo`,
-        pathname: '/sign-in/callback',
-        assign: jest.fn(),
-      },
-    });
-    /* eslint-enable @silverhand/fp/no-mutating-methods */
-
-    renderWithPageContext(
-      <SettingsProvider>
-        <MemoryRouter initialEntries={['/sign-in/callback/github']}>
-          <Routes>
-            <Route path="/sign-in/callback/:connector" element={<SecondarySocialSignIn />} />
-          </Routes>
-        </MemoryRouter>
-      </SettingsProvider>
-    );
-
-    await waitFor(() => {
-      expect(signInWithSocialSpy).toBeCalled();
-    });
   });
 });
