+++ packages/ui/src/containers/SocialSignIn/SecondarySocialSignIn/index.tsx
@@ -1,8 +1,8 @@
 import React, { useMemo, useState, useRef } from 'react';
 
+import useNativeMessageListener from '@/hooks/use-native-message-listener';
 import usePlatform from '@/hooks/use-platform';
 import useSocial from '@/hooks/use-social';
-import useSocialSignInListener from '@/hooks/use-social-signin-listener';
 
 import SocialSignInDropdown from '../SocialSignInDropdown';
 import SocialSignInIconList from '../SocialSignInIconList';
@@ -20,7 +20,7 @@ const SecondarySocialSignIn = ({ className }: Props) => {
   const moreButtonRef = useRef<HTMLButtonElement>(null);
   const { isMobile } = usePlatform();
 
-  useSocialSignInListener();
+  useNativeMessageListener();
 
   const isCollapsed = socialConnectors.length > defaultSize;
