+++ packages/ui/src/pages/SignIn/SocialCallback.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import useSocialSignInListener from '@/hooks/use-social-sign-in-listener';
+
+import SignIn from '.';
+
+const SocialCallback = () => {
+  useSocialSignInListener();
+
+  return <SignIn />;
+};
+
+export default SocialCallback;
